Se crea clase para leer datos de la tarjeta. Se agregan metodo de lectura y tambien un metodo para saber si la tarjeta es nueva

+++ app/src/main/java/com/example/vistaspruebas/TarjetaResponse.kt
@@ -23,7 +23,7 @@ data class Sectore(
     val accessBits: String,
     //@JsonProperty("_id")
     val id: String,
-    val sector: Long,
+    val sector: Int,
     val nombre: String,
     val keyA: String,
     val keyB: String,
